Sets the db relation data on start
If another charm was related to this MongoDB charm before it
finished starting, then the MongoDB would not have set the
needed relation data. In order to prevent this, the relations
had to be created after the MongoDB became active.

This should address the issue. Added a unit test to verify this
behaviour.